+++ changelog.md
@@ -2,6 +2,7 @@
 - initial work on #6: "generate conf templates."
   As part of this, some refactoring, mainly around a new Type class
   that the various accessors now use.
+- entry with no explicit type is now considered optional (with given value as default)
 
 2016-02-07 - 0.2.1
 - fix #5: incorrect reference with "config" fragment in path